Struct audioware::engine::Engine

pub struct Engine;

Audio engine built on top of kira.

Implementations

impl Engine

pub(crate) fn setup() -> Result<(), Error>

Engine setup for Manager, Tracks and Scene.

pub fn define_subtitles(package: Ref<LocalizationPackage>)

Define LocalizationPackage subtitles from Manifests.

pub fn supported_languages() -> Vec<CName>

Engine supported languages.

pub(crate) fn shutdown()

Shutdown engine.

pub fn notify(update: Lifecycle)

Notify lifecycle updates.

pub fn send(command: Command)

Send sound command (cancelable).

pub(crate) fn send_non_cancelable(command: Command)

Send non-cancelable sound command.

fn send_as(command: Command, cancelable: bool)

pub fn register_emitter( entity_id: EntityId, emitter_name: Opt<CName>, emitter_settings: Opt<EmitterSettings>, ) -> bool

Register an audio emitter to spatial Scene.

⚠️ Returns true if already registered.

pub fn unregister_emitter(entity_id: EntityId) -> bool

Unregister an audio emitter from spatial Scene.

⚠️ Returns true if already unregistered (or never registered).

pub fn is_registered_emitter(entity_id: EntityId) -> bool

pub fn emitters_count() -> i32

Current number of registered Scene audio emitters.

pub(crate) fn on_emitter_dies(entity_id: EntityId)

Whenever Scene audio emitter dies in-game.

pub(crate) fn toggle_sync_emitters(enable: bool)

Toggle Scene audio emitters synchonization.

pub(crate) fn should_sync_emitters() -> bool

Whether Scene audio emitters should be synchronized or not.

pub(crate) fn sync_emitters()

Scene audio emitters synchronization.

pub(crate) fn sync_listener()

Scene audio listener synchronization.

pub(crate) fn reclaim()

Free Manager storage from stopped sounds.

fn play( sound_name: CName, entity_id: Opt<EntityId>, emitter_name: Opt<CName>, line_type: Opt<ScnDialogLineType>, tween: Ref<Tween>, )

Play sound with optional tween.

fn play_with( sound_name: CName, entity_id: Opt<EntityId>, emitter_name: Opt<CName>, line_type: Opt<ScnDialogLineType>, ext: Ref<AudioSettingsExt>, )

Play sound with alternate settings.

fn stop( event_name: CName, entity_id: Opt<EntityId>, emitter_name: Opt<CName>, tween: Ref<Tween>, )

Stop sound with optional tween.

fn pause(tween: Ref<Tween>)

Pause all sounds with optional tween.

fn resume(tween: Ref<Tween>)

Resume all sounds with optional tween, except those already stopped.

fn switch( switch_name: CName, switch_value: CName, entity_id: Opt<EntityId>, emitter_name: Opt<CName>, switch_name_tween: Ref<Tween>, switch_value_settings: Ref<AudioSettingsExt>, )

Switch one sound for another, with optional tween and alternate settings.

fn play_on_emitter( sound_name: CName, entity_id: EntityId, emitter_name: CName, tween: Ref<Tween>, )

Play sound on audio emitter with optional tween.

fn stop_on_emitter( event_name: CName, entity_id: EntityId, emitter_name: CName, tween: Ref<Tween>, )

Stop sound on audio emitter with optional tween.

fn stop_for(entity_id: EntityId)

Stop any sound for given EntityId.

fn set_reverb_mix(value: f32)

Set ReverbMix.

fn set_preset(value: Preset)

Set audio Preset.

fn set_volume(setting: CName, value: f64)

fn context( sound_name: &CName, ) -> Result<(MutexGuard<'_, AudioManager>, SpokenLocale, WrittenLocale, Option<PlayerGender>, &'_ Id), Error>

Retrieve current engine context.