Module audioware::types::audioware::tweens

Structs

ElasticTween

Intermediate representation for kira::tween::Tween used in Redscript.

LinearTween

Intermediate representation for kira::tween::Tween used in Redscript.

Tween

Intermediate representation for kira::tween::Tween.

Traits

ToEasing

Any type that can be converted into an optional kira::tween::Easing.

ToTween

Any type that can converted into an optional kira::tween::Tween.