audioware

Trait AsGameObjectExt

pub trait AsGameObjectExt {
    // Required method
    fn resolve_display_name(&self) -> String;
}

Required Methods

fn resolve_display_name(&self) -> String

Provide a simple interface to resolve the name the most likely to be displayed for any game object and descendants.